Use a different entrypoint for android builds
iOS/PC/Mac use the main function, android uses sokol_main. However, there is a stub function
in sokol for the sokol_main function when the target is iOS/PC/Mac which causes a compile issue.
So this separates the entrypoints where on the iOS/PC/Mac entrypoint (main.zig),
sokol_main is not defined, but on the android entrypoint (main.android.zig), it is.