Include patched zig lib in order to make android build work
Android bionic does not include the getcontext function in the c std library, and in the zig std library there is a compileError because of this.
This makes compiling android binaries with lib c impossible. This commit adds a dependency that patches the zig std library for android builds.

- Added zig-lib-with-patches dependency
- Changed lib/std/c.zig to not compileError getcontext for andriod. Instead has a stub.
- Override zig_lib_dir for android artifacts use overriden lib dependency

Bonus memes:
- No more noisy zipcreate
- setLibCFile on sokol lib
- Better step name